Update Main.java
Professor Greenwell, I am trying to validate the password. I am not getting any errors, other than that I think there is something wrong with my "if statement". Can you please take a look and help me with this. I know this is not on the homework, but I was trying to learn it.